refactor(sql-orm-client): extract row-scope machinery into query-plan-scope
Move createTableRefRemapper, buildStateWhere, its cursor lowering
(buildCursorWhere, createBoundaryExpr, buildLexicographicCursorWhere),
and wrapWithRowNumberDedup out of query-plan-select.ts into a new
sibling module. query-plan-select.ts imports them back; the SQL it
compiles is unchanged. This gives compileAggregate a module to import
the same row-scoping machinery from, without duplicating it.
